requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
langchain-postgres==0.0.12
langchain-text-splitters==0.3.0
numpy==1.26.4
ollama==0.3.3
pgvector==0.2.5
psycopg==3.2.2
psycopg2==2.9.9
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.38.0

